# api/openapi.yaml
openapi: 3.1.2
info:
  title: API Management Tax
  version: 0.1.0
  summary: Governança tributária multi-jurisdicional para América Latina e Caribe
  description: >-
    Contrato inicial do produto interno. O nome da organização é configuração do ambiente;
    exemplos usam dados sintéticos e não representam aconselhamento jurídico ou tributário.
  license:
    name: Proprietary
    identifier: LicenseRef-Proprietary
servers:
  - url: /
    description: Mesmo host da aplicação
...
paths:
...
  /v1/reference/logistics-tax:
    get:
      summary: Consultar classificações tributário-logísticas
      tags: [LogisticsTax]
      operationId: getLogisticsTaxReference
      responses:
        "200": { description: Modais e tipos de operação suportados }
        "401": { $ref: "#/components/responses/Unauthorized" }
...
  /v1/operational-profiles:
    get:
      summary: Listar perfis operacionais por jurisdição
      tags: [LogisticsTax]
      operationId: listOperationalProfiles
      responses:
        "200": { $ref: "#/components/responses/DomainRecordList" }
        "401": { $ref: "#/components/responses/Unauthorized" }
    post:
      summary: Criar perfil operacional
      tags: [LogisticsTax]
      operationId: createOperationalProfile
      requestBody: { $ref: "#/components/requestBodies/DomainRecord" }
      responses:
        "201": { $ref: "#/components/responses/DomainRecordCreated" }
        "400": { $ref: "#/components/responses/BadRequest" }
        "401": { $ref: "#/components/responses/Unauthorized" }
        "403": { $ref: "#/components/responses/Forbidden" }
  /v1/legal-entities:
    get:
      summary: Listar entidades legais
      tags: [LogisticsTax]
      operationId: listLegalEntities
      responses:
        "200": { $ref: "#/components/responses/DomainRecordList" }
        "401": { $ref: "#/components/responses/Unauthorized" }
    post:
      summary: Criar entidade legal anonimizada
      tags: [LogisticsTax]
      operationId: createLegalEntity
      requestBody: { $ref: "#/components/requestBodies/DomainRecord" }
      responses:
        "201": { $ref: "#/components/responses/DomainRecordCreated" }
        "400": { $ref: "#/components/responses/BadRequest" }
        "401": { $ref: "#/components/responses/Unauthorized" }
        "403": { $ref: "#/components/responses/Forbidden" }
  /v1/establishments:
    get:
      summary: Listar estabelecimentos logísticos
      tags: [LogisticsTax]
      operationId: listEstablishments
      responses:
        "200": { $ref: "#/components/responses/DomainRecordList" }
        "401": { $ref: "#/components/responses/Unauthorized" }
    post:
      summary: Criar estabelecimento logístico
      tags: [LogisticsTax]
      operationId: createEstablishment
      requestBody: { $ref: "#/components/requestBodies/DomainRecord" }
      responses:
        "201": { $ref: "#/components/responses/DomainRecordCreated" }
        "400": { $ref: "#/components/responses/BadRequest" }
        "401": { $ref: "#/components/responses/Unauthorized" }
        "403": { $ref: "#/components/responses/Forbidden" }
  /v1/logistics-lanes:
    get:
      summary: Listar rotas e fluxos logísticos
      tags: [LogisticsTax]
      operationId: listLogisticsLanes
      responses:
        "200": { $ref: "#/components/responses/DomainRecordList" }
        "401": { $ref: "#/components/responses/Unauthorized" }
    post:
      summary: Criar rota logística multimodal
      tags: [LogisticsTax]
      operationId: createLogisticsLane
      requestBody: { $ref: "#/components/requestBodies/DomainRecord" }
      responses:
        "201": { $ref: "#/components/responses/DomainRecordCreated" }
        "400": { $ref: "#/components/responses/BadRequest" }
        "401": { $ref: "#/components/responses/Unauthorized" }
        "403": { $ref: "#/components/responses/Forbidden" }
  /v1/customs-regimes:
    get:
      summary: Listar regimes aduaneiros e zonas especiais
      tags: [LogisticsTax]
      operationId: listCustomsRegimes
      responses:
        "200": { $ref: "#/components/responses/DomainRecordList" }
        "401": { $ref: "#/components/responses/Unauthorized" }
    post:
      summary: Criar regime aduaneiro sujeito a validação local
      tags: [LogisticsTax]
      operationId: createCustomsRegime
      requestBody: { $ref: "#/components/requestBodies/DomainRecord" }
      responses:
        "201": { $ref: "#/components/responses/DomainRecordCreated" }
        "400": { $ref: "#/components/responses/BadRequest" }
        "401": { $ref: "#/components/responses/Unauthorized" }
        "403": { $ref: "#/components/responses/Forbidden" }
  /v1/tax-rules:
    get:
      summary: Listar regras tributárias por operação
      tags: [LogisticsTax]
      operationId: listTaxRules
      responses:
        "200": { $ref: "#/components/responses/DomainRecordList" }
        "401": { $ref: "#/components/responses/Unauthorized" }
    post:
      summary: Criar regra tributária com fonte e status jurídico
      tags: [LogisticsTax]
      operationId: createTaxRule
      requestBody: { $ref: "#/components/requestBodies/DomainRecord" }
      responses:
        "201": { $ref: "#/components/responses/DomainRecordCreated" }
        "400": { $ref: "#/components/responses/BadRequest" }
        "401": { $ref: "#/components/responses/Unauthorized" }
        "403": { $ref: "#/components/responses/Forbidden" }
  /v1/tax-documents:
    get:
      summary: Listar documentos fiscais, de transporte e aduaneiros
      tags: [LogisticsTax]
      operationId: listTaxDocuments
      responses:
        "200": { $ref: "#/components/responses/DomainRecordList" }
        "401": { $ref: "#/components/responses/Unauthorized" }
    post:
      summary: Registrar metadados de documento tributário-logístico
      tags: [LogisticsTax]
      operationId: createTaxDocument
      requestBody: { $ref: "#/components/requestBodies/DomainRecord" }
      responses:
        "201": { $ref: "#/components/responses/DomainRecordCreated" }
        "400": { $ref: "#/components/responses/BadRequest" }
        "401": { $ref: "#/components/responses/Unauthorized" }
        "403": { $ref: "#/components/responses/Forbidden" }
  /v1/tax-recovery-opportunities:
    get:
      summary: Listar oportunidades de recuperação de tributos
      tags: [LogisticsTax]
      operationId: listTaxRecoveryOpportunities
      responses:
        "200": { $ref: "#/components/responses/DomainRecordList" }
        "401": { $ref: "#/components/responses/Unauthorized" }
    post:
      summary: Criar oportunidade e classificar risco prescricional
      tags: [LogisticsTax]
      operationId: createTaxRecoveryOpportunity
      requestBody: { $ref: "#/components/requestBodies/DomainRecord" }
      responses:
        "201": { $ref: "#/components/responses/DomainRecordCreated" }
        "400": { $ref: "#/components/responses/BadRequest" }
        "401": { $ref: "#/components/responses/Unauthorized" }
        "403": { $ref: "#/components/responses/Forbidden" }
  /v1/permanent-establishment-assessments:
    get:
      summary: Listar avaliações preliminares de estabelecimento permanente
      tags: [LogisticsTax]
      operationId: listPermanentEstablishmentAssessments
      responses:
        "200": { $ref: "#/components/responses/DomainRecordList" }
        "401": { $ref: "#/components/responses/Unauthorized" }
    post:
      summary: Criar indicador preliminar de estabelecimento permanente
      tags: [LogisticsTax]
      operationId: createPermanentEstablishmentAssessment
      requestBody: { $ref: "#/components/requestBodies/DomainRecord" }
      responses:
        "201": { $ref: "#/components/responses/DomainRecordCreated" }
        "400": { $ref: "#/components/responses/BadRequest" }
        "401": { $ref: "#/components/responses/Unauthorized" }
        "403": { $ref: "#/components/responses/Forbidden" }
  /v1/integration-connections:
    get:
      summary: Listar conexões com ERP, TMS, WMS e fontes externas
      tags: [LogisticsTax]
      operationId: listIntegrationConnections
      responses:
        "200": { $ref: "#/components/responses/DomainRecordList" }
        "401": { $ref: "#/components/responses/Unauthorized" }
    post:
      summary: Configurar conexão por referência segura de segredo
      tags: [LogisticsTax]
      operationId: createIntegrationConnection
      requestBody: { $ref: "#/components/requestBodies/DomainRecord" }
      responses:
        "201": { $ref: "#/components/responses/DomainRecordCreated" }
        "400": { $ref: "#/components/responses/BadRequest" }
        "401": { $ref: "#/components/responses/Unauthorized" }
        "403": { $ref: "#/components/responses/Forbidden" }
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: Em produção, token do provedor corporativo. Cabeçalhos sintéticos existem apenas localmente.
...
  requestBodies:
    DomainRecord:
      required: true
      description: >-
        Payload específico do domínio, validado estritamente pela API. Campos legais devem
        incluir fonte e status de validação quando aplicável.
      content:
        application/json:
          schema:
            type: object
            minProperties: 1
            additionalProperties: true
  responses:
    DomainRecordList:
      description: Registros visíveis no tenant e nos escopos de país do ator
      content:
        application/json:
          schema:
            type: object
            required: [data]
            properties:
              data:
                type: array
                items: { $ref: "#/components/schemas/DomainRecord" }
    DomainRecordCreated:
      description: Registro criado e evento de auditoria encadeado
      content:
        application/json:
          schema: { $ref: "#/components/schemas/DomainRecord" }
    BadRequest:
      description: Requisição inválida
      content:
        application/problem+json:
          schema: { $ref: "#/components/schemas/Problem" }
    Unauthorized:
      description: Identidade ausente ou inválida
      content:
        application/problem+json:
          schema: { $ref: "#/components/schemas/Problem" }
    Forbidden:
      description: Identidade válida sem permissão efetiva
      content:
        application/problem+json:
          schema: { $ref: "#/components/schemas/Problem" }
...
  schemas:
...
    DomainRecord:
      type: object
      required: [id, tenantId, createdAt, updatedAt]
      properties:
        id: { type: string, format: uuid }
        tenantId: { type: string }
        createdAt: { type: string, format: date-time }
        updatedAt: { type: string, format: date-time }
        legalValidationStatus:
          type: string
          enum: [PRELIMINARY, PENDING_LOCAL_COUNSEL, VALIDATED, SUPERSEDED]
      additionalProperties: true
...
    Problem:
      type: object
      required: [type, title, status, detail, instance, correlationId]
      properties:
        type: { type: string, format: uri }
        title: { type: string }
        status: { type: integer, minimum: 400, maximum: 599 }
        detail: { type: string }
        instance: { type: string }
        correlationId: { type: string }
security:
  - bearerAuth: []